02. oliveyoung search 'menscare' - claud/requirements.txt
playwright>=1.44.0
requests>=2.31.0
PyNaCl>=1.5.0
